Refactor conftest.py into focused helper modules
The conftest.py file grew to over 1200 lines, mixing unrelated
concerns such as setup checks, pre-test routines, log collection,
and Windows monitoring. This made it difficult to navigate and
maintain.

Extract four dedicated helper modules from conftest.py:
- helpers_setup_checks.py: environment validation fixtures
- helpers_pretest.py: pre-test preparation logic
- helpers_log_collection.py: test log gathering utilities
- helpers_windows_monitoring.py: Windows VM monitoring helpers

conftest.py now delegates to these modules, reducing it to ~254
lines and giving each concern a clear, single-responsibility home.
