+++ nat-lab/tests/conftest.py
@@ -1,61 +1,31 @@
-# TODO (LLT-7084): Move fixtures to separate fixture files out of conftest.py
-# pylint: disable=too-many-lines
 import asyncio
-import base64
-import itertools
-import json
 import logging
 import os
 import Pyro5  # type: ignore
 import pytest
-import re
-import shutil
-import ssl
-import subprocess
-import tests.config
 import threading
-import time
-import urllib.error
-import urllib.request
-from collections import defaultdict
 from contextlib import AsyncExitStack
-from datetime import datetime
-from http import HTTPStatus
-from itertools import combinations
-from tests.config import LAN_ADDR_MAP, CORE_API_IP, CORE_API_CREDENTIALS
 from tests.helpers import SetupParameters
-from tests.interderp_cli import InterDerpClient
+from tests.helpers_log_collection import collect_logs, collect_kernel_logs
+from tests.helpers_pretest import perform_pretest_cleanups, copy_vm_binaries_if_needed
+from tests.helpers_setup_checks import (
+    perform_setup_checks,
+    check_gateway_connectivity,
+    check_all_containers_running,
+    get_session_vm_marks,
+)
+from tests.helpers_windows_monitoring import (
+    start_tcpdump_processes,
+    start_windows_vms_resource_monitoring,
+)
 from tests.log_collector import LOG_COLLECTORS
 from tests.utils.bindings import TelioAdapterType
-from tests.utils.connection import ConnectionTag, TargetOS, clear_ephemeral_setups_set
-from tests.utils.connection.docker_connection import DockerConnection, container_id
-from tests.utils.connection.ssh_connection import SshConnection
-from tests.utils.connection_util import new_connection_raw, is_running
-from tests.utils.logger import log, setup_log
-from tests.utils.process import ProcessExecError
-from tests.utils.router import IPStack, new_router
-from tests.utils.tcpdump import make_local_tcpdump, make_tcpdump
+from tests.utils.connection import ConnectionTag
+from tests.utils.connection_util import new_connection_raw
+from tests.utils.logger import log
+from tests.utils.router import IPStack
 from tests.utils.testing import get_current_test_log_path
 from typing import List
-
-DERP_SERVER_1_ADDR = "http://10.0.10.1:8765"
-DERP_SERVER_2_ADDR = "http://10.0.10.2:8765"
-DERP_SERVER_3_ADDR = "http://10.0.10.3:8765"
-DERP_SERVER_1_SECRET_KEY = "yBTYHj8yPlG9VtMYMwJSRHdzNdyAlVXGc6X2xJkjfHQ="
-DERP_SERVER_2_SECRET_KEY = "2NgALOCSKJcDxwr8MtA+6lYbf7b98KSdAROGoUwZ1V0="
-
-SETUP_CHECK_TIMEOUT_S = 30
-SETUP_CHECK_RETRIES = 5
-SETUP_CHECK_CONNECTIVITY_TIMEOUT = 60
-SETUP_CHECK_CONNECTIVITY_RETRIES = 1
-GW_CHECK_CONNECTIVITY_TIMEOUT = 30
-GW_CHECK_CONNECTIVITY_RETRIES = 2
-SETUP_CHECK_MAC_COLLISION_TIMEOUT_S = 300
-SETUP_CHECK_MAC_COLLISION_RETRIES = 1
-SETUP_CHECK_ARP_CACHE_TIMEOUT_S = 300
-SETUP_CHECK_ARP_CACHE_RETRIES = 1
-SETUP_CHECK_DUPLICATE_IP_TIMEOUT_S = 60
-SETUP_CHECK_DUPLICATE_IP_RETRIES = 1
 
 RUNNER: asyncio.Runner | None = None
 SESSION_SCOPE_EXIT_STACK: AsyncExitStack | None = None
@@ -63,21 +33,8 @@ TASKS: List[asyncio.Task] = []
 END_TASKS: threading.Event = threading.Event()
 CURRENT_TEST_LOG_FILE = None
 
-LOG_DIR = "logs"
-OPENWRT_VM_TAGS = [ConnectionTag.VM_OPENWRT_GW_1, ConnectionTag.VM_OPENWRT_GW_2]
 SESSION_VM_MARKS: set[str] = set()
 SESSION_IS_CONTAINER_RUNNING: dict[ConnectionTag, bool] = {}
-SESSION_MARK_TO_CONTAINERS = {
-    "fullcone": [
-        ConnectionTag.VM_LINUX_FULLCONE_GW_1,
-        ConnectionTag.VM_LINUX_FULLCONE_GW_2,
-    ],
-    "mac": [ConnectionTag.VM_MAC],
-    "nlx": [ConnectionTag.VM_LINUX_NLX_1],
-    "openwrt": OPENWRT_VM_TAGS,
-    "windows": [ConnectionTag.VM_WINDOWS_1],
-    "windows2": [ConnectionTag.VM_WINDOWS_2],
-}
 
 
 def _cancel_all_tasks(loop: asyncio.AbstractEventLoop):
@@ -156,625 +113,6 @@ def pytest_make_parametrize_id(config, val):
     return param_id
 
 
-async def setup_check_interderp():
-    setup_log.info("Running interderp checks..")
-    async with AsyncExitStack() as exit_stack:
-        connections = [
-            await exit_stack.enter_async_context(new_connection_raw(conn_tag))
-            for conn_tag in [
-                ConnectionTag.DOCKER_CONE_CLIENT_1,
-                ConnectionTag.DOCKER_DERP_1,
-                ConnectionTag.DOCKER_DERP_2,
-                ConnectionTag.DOCKER_DERP_3,
-            ]
-        ]
-
-        if not isinstance(connections[0], DockerConnection):
-            raise Exception("Not docker connection")
-
-        async with make_tcpdump(connections):
-            for idx, (server1, server2) in enumerate(
-                combinations(
-                    [DERP_SERVER_1_ADDR, DERP_SERVER_2_ADDR, DERP_SERVER_3_ADDR], 2
-                )
-            ):
-                derp_test = InterDerpClient(
-                    connections[0],
-                    server1,
-                    server2,
-                    DERP_SERVER_1_SECRET_KEY,
-                    DERP_SERVER_2_SECRET_KEY,
-                    idx,
-                )
-                await derp_test.execute()
-                await derp_test.save_logs()
-
-
-async def setup_check_duplicate_ip_addresses():
-    """
-    Enumerate all running Docker containers, gather their IPv4 addresses (excluding 127.0.0.1),
-    and fail if any IPv4 address is used by more than one container.
-    Converted from nat-lab/check_ip.sh.
-    """
-    setup_log.info("Running duplicate IP addresses check..")
-    try:
-        containers_out = subprocess.check_output(
-            ["docker", "ps", "-q"], text=True
-        ).strip()
-    except Exception as e:  # pylint: disable=broad-exception-caught
-        setup_log.error(
-            f"cannot execute 'docker ps -q': {e}; skipping duplicate IP check"
-        )
-        return
-
-    if not containers_out:
-        setup_log.error("No running containers found.")
-        return
-
-    containers = [c for c in containers_out.splitlines() if c.strip()]
-
-    ip_owner: dict[str, str] = {}
-    duplicates: dict[str, set[str]] = defaultdict(set)
-
-    for cid in containers:
-        try:
-            name = subprocess.check_output(
-                ["docker", "inspect", "--format={{.Name}}", cid],
-                text=True,
-            ).strip()
-            name = re.sub(r"^/+", "", name)
-        except subprocess.CalledProcessError:
-            name = cid
-
-        # Extract IPv4 addresses inside the container without relying on grep -P.
-        # Use: ip -4 -o addr show -> "... IFACE ... A.B.C.D/XX ..."
-        # Then project the CIDR column and strip mask.
-        try:
-            ips_out = subprocess.check_output(
-                [
-                    "docker",
-                    "exec",
-                    cid,
-                    "sh",
-                    "-c",
-                    "ip -4 -o addr show | awk '{print $4}' | cut -d/ -f1",
-                ],
-                text=True,
-            )
-        except subprocess.CalledProcessError:
-            ips_out = ""
-
-        ips = [
-            ip.strip()
-            for ip in ips_out.split()
-            if ip.strip() and not ip.startswith("127.")
-        ]
-
-        # Print container name and IPs (for debugging parity with the original script)
-        setup_log.debug("========== %s (%s) ==========", name, cid)
-        if not ips:
-            setup_log.debug("No IPs found")
-        else:
-            for ip in ips:
-                setup_log.debug(f"IP: {ip}")
-
-        # Detect duplicates
-        for ip in ips:
-            # 100.64.0.1 is a libtelio hardcoded address representing VPN virtual peer,
-            # common for every client
-            if ip in ip_owner and ip_owner[ip] != name and ip != "100.64.0.1":
-                duplicates[ip].update({ip_owner[ip], name})
-            else:
-                ip_owner[ip] = name
-
-    if duplicates:
-        for ip, owners in sorted(duplicates.items()):
-            setup_log.warning(
-                "  -> Duplicate IP %s found! Used by containers: %s",
-                ip,
-                ", ".join(sorted(owners)),
-            )
-        details = {ip: sorted(list(owners)) for ip, owners in duplicates.items()}
-        raise Exception(f"Found duplicate container IPv4 addresses: {details}")
-
-
-async def setup_check_duplicate_mac_addresses():
-    setup_log.info("Running duplicate MAC addresses check..")
-    mac_re = re.compile(r"(?:[0-9a-f]{2}[:\-]){5}[0-9a-f]{2}", re.IGNORECASE)
-    seen = defaultdict(set)  # mac -> set of ConnectionTag
-
-    ignore_macs = {
-        "00:00:00:00:00:00",
-        "ff:ff:ff:ff:ff:ff",
-    }
-
-    async with AsyncExitStack() as exit_stack:
-        for conn_tag in ConnectionTag:
-            if not SESSION_IS_CONTAINER_RUNNING[conn_tag]:
-                setup_log.debug(
-                    "%s is not running, skipping duplicate MAC address check..",
-                    conn_tag.name,
-                )
-                continue
-
-            try:
-                conn = await exit_stack.enter_async_context(
-                    new_connection_raw(conn_tag)
-                )
-            except Exception as e:  # pylint: disable=broad-exception-caught
-                setup_log.warning(
-                    "Failed to check MAC address for %s: %s", conn_tag.name, e
-                )
-                raise e
-
-            if conn.target_os == TargetOS.Linux:
-                cmd = ["sh", "-c", "ip link show | awk '/link\\/ether/ {print $2}'"]
-            elif conn.target_os == TargetOS.Mac:
-                cmd = ["sh", "-c", "ifconfig | awk '/ether/ {print $2}'"]
-            elif conn.target_os == TargetOS.Windows:
-                cmd = ["getmac", "/v", "/fo", "list"]
-            else:
-                raise Exception("unknown target os")
-
-            proc = await conn.create_process(cmd, quiet=True).execute()
-            output = proc.get_stdout()
-
-            # extract MACs (handles both ":" and "-" formats)
-            for m in mac_re.finditer(output):
-                normalized = m.group(0).lower().replace("-", ":")
-                if normalized in ignore_macs:
-                    continue
-                seen[normalized].add(conn_tag)
-
-    duplicates = {
-        mac: sorted(map(str, tags)) for mac, tags in seen.items() if len(tags) > 1
-    }
-    if duplicates:
-        for mac, tags in duplicates.items():
-            setup_log.error("%s -> %s", mac, ", ".join(tags))
-        raise Exception(f"Found duplicate MACs: {duplicates}")
-
-
-async def setup_check_arp_cache():
-    """
-    Ensure all VM LAN_ADDR_MAP IPv4 addresses are present in the host ARP cache
-    and are in a usable state.
-    """
-    if TargetOS.local() != TargetOS.Linux:
-        setup_log.info("setup_check: skipping ARP cache validation on non-Linux host")
-        return
-
-    setup_log.info("Running ARP cache check..")
-
-    def warm_arp(ip: str) -> None:
-        subprocess.call(
-            ["ping", "-c", "1", "-W", "1", ip],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-
-    def read_arp_entries() -> list[dict]:
-        return json.loads(
-            subprocess.check_output(
-                ["ip", "-j", "neigh", "show"],
-                text=True,
-            ).strip()
-        )
-
-    subprocess.call(
-        ["sudo", "ip", "-s", "-s", "neigh", "flush", "all"],
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.DEVNULL,
-    )
-
-    acceptable_states = {"REACHABLE", "STALE", "DELAY", "PROBE", "PERMANENT"}
-    failures: list[str] = []
-
-    for tag in get_required_vm_containers_from_marks():
-        if tag in OPENWRT_VM_TAGS:
-            continue
-        for ip in LAN_ADDR_MAP[tag].values():
-            success = False
-            last_arp_entries: list[dict] = []
-            if ip == "":
-                continue
-            while True:
-                if success:
-                    break
-                warm_arp(ip)
-                last_arp_entries = read_arp_entries()
-                for e in last_arp_entries:
-                    dst_ip = e.get("dst")
-                    lladdr = e.get("lladdr")
-                    state = e.get("state")
-                    if dst_ip is None or dst_ip != ip:
-                        continue
-                    if lladdr is None:
-                        continue
-                    if state is None or state[0] not in acceptable_states:
-                        continue
-                    success = True
-                    break
-            if not success:
-                state = next(
-                    (
-                        e.get("state", "missing")
-                        for e in last_arp_entries
-                        if e.get("dst") == ip
-                    ),
-                    "missing",
-                )
-                failures.append(f"{tag.name}:{ip} state={state}")
-
-    if failures:
-        raise Exception("ARP cache not ready for VMs: " + ", ".join(failures))
-
-
-async def setup_nlx_vpn_server() -> None:
-    """Populate NLX_SERVER keys in tests.config (global, survives between tests)."""
-    async with AsyncExitStack() as exit_stack:
-        conn = await exit_stack.enter_async_context(
-            new_connection_raw(ConnectionTag.VM_LINUX_NLX_1)
-        )
-        nlx_server = tests.config.NLX_SERVER
-        container = nlx_server.get("container")
-
-        get_pub_cmd = 'nlx | awk \'$1=="public" && $2=="key:" {print $3; exit}\''
-        proc = await conn.create_process(["bash", "-lc", get_pub_cmd]).execute()
-        pub_key = proc.get_stdout().strip()
-
-        if not pub_key:
-            raise RuntimeError(
-                f"Could not obtain NordLynx public key from nlx on {container}"
-            )
-        nlx_server["public_key"] = pub_key
-        log.debug(
-            "NordLynx public key for %s: %s",
-            container,
-            pub_key,
-        )
-
-        get_priv_cmd = (
-            "nlx showconf nlx0 | "
-            'awk \'$1=="PrivateKey" && $2=="=" {print $3; exit}\''
-        )
-
-        proc_priv = await conn.create_process(["bash", "-lc", get_priv_cmd]).execute()
-        priv_key = proc_priv.get_stdout().strip()
-
-        if not priv_key:
-            raise RuntimeError(
-                f"Could not obtain NordLynx private key from nlx showconf on {container}"
-            )
-
-        nlx_server["private_key"] = priv_key
-
-
-SETUP_CHECKS = [
-    (
-        setup_check_duplicate_ip_addresses,
-        SETUP_CHECK_DUPLICATE_IP_TIMEOUT_S,
-        SETUP_CHECK_DUPLICATE_IP_RETRIES,
-    ),
-    (
-        setup_check_arp_cache,
-        SETUP_CHECK_ARP_CACHE_TIMEOUT_S,
-        SETUP_CHECK_ARP_CACHE_RETRIES,
-    ),
-    (
-        setup_check_duplicate_mac_addresses,
-        SETUP_CHECK_MAC_COLLISION_TIMEOUT_S,
-        SETUP_CHECK_MAC_COLLISION_RETRIES,
-    ),
-    (setup_check_interderp, SETUP_CHECK_TIMEOUT_S, SETUP_CHECK_RETRIES),
-    (setup_nlx_vpn_server, SETUP_CHECK_TIMEOUT_S, SETUP_CHECK_RETRIES),
-]
-
-
-async def perform_setup_checks() -> bool:
-    for target, timeout, retries in SETUP_CHECKS:
-        while retries > 0:
-            try:
-                await asyncio.wait_for(asyncio.shield(target()), timeout)
-                break
-            except asyncio.TimeoutError:
-                setup_log.warning("%s() timeout, retrying...", target)
-            except ProcessExecError as e:
-                setup_log.warning("%s() process exec error %s, retrying...", target, e)
-            retries -= 1
-        else:
-            return False
-
-    return True
-
-
-async def check_gateway_connectivity() -> bool:
-    setup_log.info("Checking gateways connectivity..")
-    if SESSION_SCOPE_EXIT_STACK is None:
-        raise RuntimeError("SESSION_SCOPE_EXIT_STACK is not initialized")
-    current_gateway = None
-    for _ in range(GW_CHECK_CONNECTIVITY_RETRIES + 1):
-        try:
-            for gw_tag in ConnectionTag:
-                if "_GW" in gw_tag.name:
-                    if not await is_running(gw_tag):
-                        continue
-                    current_gateway = gw_tag
-                    await SESSION_SCOPE_EXIT_STACK.enter_async_context(
-                        new_connection_raw(gw_tag)
-                    )
-            return True
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            gw_name = getattr(current_gateway, "name", "unknown")
-            setup_log.error("Failed to connect to %s", gw_name)
-            setup_log.error("Exception error: %s", e)
-            await asyncio.sleep(GW_CHECK_CONNECTIVITY_TIMEOUT)
-    # ignore connection failure in case of OpenWrt Gateway
-    if current_gateway and current_gateway in OPENWRT_VM_TAGS:
-        return True
-    return False
-
-
-async def kill_natlab_processes():
-    # Do not execute cleanup in non-CI environment,
-    # because cleanup requires elevated privileges,
-    # and it is rather inconvenient when every run
-    # requires to enter sudo password
-    # If someone really wants cleanup - one can set
-    # ENABLE_NATLAB_PROCESS_CLEANUP envron variable
-    if (
-        not "GITLAB_CI" in os.environ
-        and not "ENABLE_NATLAB_PROCESS_CLEANUP" in os.environ
-    ):
-        return
-
-    cleanup_script_path = os.path.join(
-        os.path.dirname(__file__), "../bin/cleanup_natlab_processes"
-    )
-    subprocess.run(["sudo", cleanup_script_path]).check_returncode()
-
-
-async def reset_service_credentials_cache():
-    ssl_context = ssl.create_default_context()
-    ssl_context.check_hostname = False
-    ssl_context.verify_mode = ssl.CERT_NONE
-    credentials = (
-        f"{CORE_API_CREDENTIALS['username']}:{CORE_API_CREDENTIALS['password']}"
-    )
-    encoded_credentials = base64.b64encode(credentials.encode()).decode()
-    headers = {
-        "Authorization": f"Basic {encoded_credentials}",
-        "Content-Type": "application/json",
-        "Content-Length": "0",
-    }
-    request = urllib.request.Request(
-        f"https://{CORE_API_IP}/test/reset-credentials",
-        data=b"",
-        method="POST",
-        headers=headers,
-    )
-    try:
-        with urllib.request.urlopen(request, context=ssl_context) as response:
-            if response.status == HTTPStatus.OK:
-                setup_log.debug("Service credentials cache reset successfully")
-            else:
-                setup_log.warning(
-                    "Failed to reset service credentials cache: HTTP %s",
-                    response.status,
-                )
-    except urllib.error.HTTPError as e:
-        setup_log.warning(
-            "Failed to reset service credentials cache: HTTP %s - %s", e.code, e.reason
-        )
-        raise
-    except Exception as e:  # pylint: disable=broad-exception-caught
-        setup_log.warning("Error resetting service credentials cache: %s", e)
-        raise
-
-
-async def reset_upnpd_on_upnp_gateways():
-    setup_log.info("Resetting miniupnpd on UPNP gateways..")
-    async with AsyncExitStack() as exit_stack:
-        for gw_tag in ConnectionTag:
-            if "UPNP_GW" not in gw_tag.name:
-                continue
-            try:
-                if not await is_running(gw_tag):
-                    continue
-                conn = await exit_stack.enter_async_context(new_connection_raw(gw_tag))
-                router = new_router(conn, IPStack.IPv4v6)
-                # Briefly stop and restart miniupnpd to purge any stale mappings
-                async with router.reset_upnpd():
-                    pass
-                setup_log.debug("miniupnpd reset on %s", gw_tag.name)
-            except Exception as e:  # pylint: disable=broad-exception-caught
-                setup_log.warning("Failed to reset miniupnpd on %s: %s", gw_tag.name, e)
-
-
-async def reset_machines_cache():
-    ssl_context = ssl.create_default_context()
-    ssl_context.check_hostname = False
-    ssl_context.verify_mode = ssl.CERT_NONE
-    credentials = (
-        f"{CORE_API_CREDENTIALS['username']}:{CORE_API_CREDENTIALS['password']}"
-    )
-    encoded_credentials = base64.b64encode(credentials.encode()).decode()
-    headers = {
-        "Authorization": f"Basic {encoded_credentials}",
-        "Content-Type": "application/json",
-        "Content-Length": "0",
-    }
-    request = urllib.request.Request(
-        f"https://{CORE_API_IP}/test/reset-machines",
-        data=b"",
-        method="POST",
-        headers=headers,
-    )
-    try:
-        with urllib.request.urlopen(
-            request, context=ssl_context, timeout=5
-        ) as response:
-            if response.status == HTTPStatus.OK:
-                log.debug("Machines cache reset successfully")
-            else:
-                log.warning(
-                    "Failed to reset machines cache: HTTP %s",
-                    response.status,
-                )
-    except urllib.error.HTTPError as e:
-        log.debug("Failed to reset machines cache: HTTP %s - %s", e.code, e.reason)
-    except Exception as e:  # pylint: disable=broad-exception-caught
-        log.debug("Could not reset machines cache (server may not be running): %s", e)
-
-
-PRETEST_CLEANUPS = [
-    kill_natlab_processes,
-    clear_ephemeral_setups_set,
-    reset_service_credentials_cache,
-    reset_upnpd_on_upnp_gateways,
-    reset_machines_cache,
-]
-
-
-async def perform_pretest_cleanups():
-    for cleanup in PRETEST_CLEANUPS:
-        await cleanup()
-
-
-async def _copy_vm_binaries(tag: ConnectionTag):
-    try:
-        setup_log.info("Copying binaries for %s", tag)
-        async with SshConnection.new_connection(
-            LAN_ADDR_MAP[tag]["primary"], tag, copy_binaries=True
-        ):
-            pass
-    except OSError as e:
-        setup_log.error(e)
-        raise e
-
-
-async def _copy_vm_binaries_if_needed():
-    if "windows" in SESSION_VM_MARKS:
-        await _copy_vm_binaries(ConnectionTag.VM_WINDOWS_1)
-        try:
-            await _copy_vm_binaries(ConnectionTag.VM_WINDOWS_2)
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            setup_log.warning("[Ignored] Couldn't copy binary to VM_WINDOWS_2: %s", e)
-    if "mac" in SESSION_VM_MARKS:
-        await _copy_vm_binaries(ConnectionTag.VM_MAC)
-    if "openwrt" in SESSION_VM_MARKS:
-        await _copy_vm_binaries(ConnectionTag.VM_OPENWRT_GW_1)
-        await _copy_vm_binaries(ConnectionTag.VM_OPENWRT_GW_2)
-
-
-def save_dmesg_from_host(suffix):
-    try:
-        result = subprocess.run(
-            ["sudo", "dmesg", "-d", "-T"],
-            capture_output=True,
-            text=True,
-            check=True,
-        ).stdout
-    except subprocess.CalledProcessError as e:
-        setup_log.error("Error executing dmesg: %s", e)
-        return
-
-    if result:
-        with open(
-            os.path.join(LOG_DIR, f"dmesg-{suffix}.txt"), "w", encoding="utf-8"
-        ) as f:
-            f.write(result)
-
-
-async def save_dmesg_from_remote_vm(conn_tag: ConnectionTag, suffix: str) -> None:
-    os.makedirs(LOG_DIR, exist_ok=True)
-
-    file_suffix = f"{conn_tag.name.lower()}-{suffix}"
-    log_path = os.path.join(LOG_DIR, f"dmesg-{file_suffix}.txt")
-
-    async with new_connection_raw(conn_tag) as conn:
-        dmesg_cmd = ["dmesg", "-d", "-T"]
-        try:
-            proc = await conn.create_process(dmesg_cmd, quiet=True).execute()
-            stdout = proc.get_stdout() or ""
-            with open(log_path, "w", encoding="utf-8") as f:
-                f.write(stdout)
-        except ProcessExecError as e:
-            setup_log.warning(
-                "Failed to collect remote dmesg from %s. Return code=%s, stderr=%r, stdout=%r",
-                conn_tag,
-                e.returncode,
-                e.stderr,
-                e.stdout,
-            )
-
-
-def save_audit_log_from_host(suffix):
-    try:
-        source_path = "/var/log/audit/audit.log"
-        if os.path.exists(source_path):
-            shutil.copy2(source_path, f"{LOG_DIR}/audit_{suffix}.log")
-        else:
-            setup_log.warning("The audit file %s", source_path)
-    except Exception as e:  # pylint: disable=broad-exception-caught
-        setup_log.warning("An error occurred when processing audit log: %s", e)
-
-
-async def save_nordlynx_logs():
-    if "nlx" not in SESSION_VM_MARKS:
-        return
-
-    source_log_dir_path = "/var/log"
-    nlx_log_files = [
-        "nlx-radius.log",
-        "pq-upgrader.log",
-        "fakefm.log",
-        "nlx-ns.log",
-        "dynamic_api_fakefm.log",
-    ]
-
-    async with new_connection_raw(ConnectionTag.VM_LINUX_NLX_1) as conn:
-        for log_file in nlx_log_files:
-            remote_path = os.path.join(source_log_dir_path, log_file)
-            local_path = os.path.join(LOG_DIR, log_file)
-            try:
-                await conn.download(remote_path, local_path)
-                log.info("Downloaded '%s' to '%s'", remote_path, local_path)
-            except Exception as e:  # pylint: disable=broad-exception-caught
-                setup_log.warning(
-                    "An error occurred when processing %s log: %s", remote_path, e
-                )
-
-
-async def _save_macos_logs(conn, suffix):
-    try:
-        dmesg_proc = await conn.create_process(["dmesg"], quiet=True).execute()
-        with open(
-            os.path.join(LOG_DIR, f"dmesg-macos-{suffix}.txt"), "w", encoding="utf-8"
-        ) as f:
-            f.write(dmesg_proc.get_stdout())
-    except ProcessExecError as e:
-        setup_log.warning("Failed to collect dmesg logs %s", e)
-
-
-async def collect_kernel_logs(suffix):
-    os.makedirs(LOG_DIR, exist_ok=True)
-
-    save_dmesg_from_host(suffix)
-    save_audit_log_from_host(suffix)
-    await save_dmesg_from_remote_vm(ConnectionTag.VM_LINUX_NLX_1, suffix)
-
-    if "mac" in SESSION_VM_MARKS:
-        try:
-            async with SshConnection.new_connection(
-                LAN_ADDR_MAP[ConnectionTag.VM_MAC]["primary"], ConnectionTag.VM_MAC
-            ) as conn:
-                await _save_macos_logs(conn, suffix)
-        except OSError as e:
-            if os.environ.get("GITLAB_CI"):
-                raise e
-
-
 def pytest_collection_modifyitems(items):
     for item in items:
         # Apply 5 minutes timeout to windows tests (due to constant lag)
@@ -782,58 +120,31 @@ def pytest_collection_modifyitems(items):
             item.add_marker(pytest.mark.timeout(300))
 
 
-def get_session_vm_marks(items):
-    global SESSION_VM_MARKS
-
-    SESSION_VM_MARKS = set()
-    for item in items:
-        for mark in item.own_markers:
-            SESSION_VM_MARKS.add(mark.name)
-
-
-async def check_all_containers_running():
-    global SESSION_IS_CONTAINER_RUNNING
-    SESSION_IS_CONTAINER_RUNNING = {}
-
-    setup_log.info("Checking running containers..")
-
-    tags = list(ConnectionTag)
-    is_running_results = await asyncio.gather(
-        *[is_running(conn_tag) for conn_tag in tags]
-    )
-
-    for conn_tag, is_running_res in zip(tags, is_running_results):
-        SESSION_IS_CONTAINER_RUNNING[conn_tag] = is_running_res
-
-    running = [
-        tag.name for tag, status in SESSION_IS_CONTAINER_RUNNING.items() if status
-    ]
-    setup_log.debug("Running (%s): %s", len(running), ", ".join(running).lower())
-
-    not_running = [
-        tag.name for tag, status in SESSION_IS_CONTAINER_RUNNING.items() if not status
-    ]
-    setup_log.debug(
-        "Not running (%s): %s", len(not_running), ", ".join(not_running).lower()
-    )
-
-
 def pytest_runtestloop(session):
+    global SESSION_VM_MARKS, SESSION_IS_CONTAINER_RUNNING
+
     if session.config.option.collectonly:
         return
 
-    get_session_vm_marks(session.items)
+    SESSION_VM_MARKS = get_session_vm_marks(session.items)
 
     if "NATLAB_SKIP_SETUP_CHECKS" not in os.environ:
-        asyncio.run(check_all_containers_running())
+        SESSION_IS_CONTAINER_RUNNING = asyncio.run(check_all_containers_running())
 
-        if not asyncio.run(perform_setup_checks()):
+        if not asyncio.run(
+            perform_setup_checks(SESSION_IS_CONTAINER_RUNNING, SESSION_VM_MARKS)
+        ):
             pytest.exit("Setup checks failed, exiting ...")
 
     if "NATLAB_SAVE_LOGS" in os.environ:
-        asyncio.run(collect_kernel_logs("before_tests"))
+        asyncio.run(
+            collect_kernel_logs(
+                "before_tests",
+                SESSION_VM_MARKS,
+            )
+        )
 
-    asyncio.run(_copy_vm_binaries_if_needed())
+    asyncio.run(copy_vm_binaries_if_needed(SESSION_VM_MARKS))
 
 
 def pytest_runtest_setup():
@@ -893,10 +204,14 @@ def pytest_sessionstart(session):
     if os.environ.get("NATLAB_SAVE_LOGS"):
         RUNNER = asyncio.Runner()
         SESSION_SCOPE_EXIT_STACK = AsyncExitStack()
-        if not RUNNER.run(check_gateway_connectivity()):
+        if not RUNNER.run(check_gateway_connectivity(SESSION_SCOPE_EXIT_STACK)):
             pytest.exit("Gateway nodes connectivity check failed, exiting ...")
-        RUNNER.run(start_tcpdump_processes())
-        RUNNER.run(start_windows_vms_resource_monitoring())
+        RUNNER.run(
+            start_tcpdump_processes(
+                SESSION_SCOPE_EXIT_STACK,
+            )
+        )
+        RUNNER.run(start_windows_vms_resource_monitoring(TASKS, END_TASKS))
 
 
 # pylint: disable=unused-argument
@@ -913,270 +228,7 @@ def pytest_sessionfinish(session, exitstatus):
         finally:
             RUNNER.close()
 
-    asyncio.run(collect_logs())
-
-
-async def collect_logs():
-    collect_nordderper_logs()
-    collect_dns_server_logs()
-    collect_core_api_server_logs()
-    await collect_kernel_logs("after_tests")
-    await collect_mac_diagnostic_reports()
-    await save_nordlynx_logs()
-
-
-def collect_nordderper_logs():
-    num_containers = 3
-
-    for i in range(1, num_containers + 1):
-        container_name = f"nat-lab-derp-{i:02d}-1"
-        destination_path = f"{LOG_DIR}/derp_{i:02d}_relay.log"
-
-        copy_file_from_container(
-            container_name, "/etc/nordderper/relay.log", destination_path
-        )
-
-
-def collect_dns_server_logs():
-    num_containers = 2
-
-    for i in range(1, num_containers + 1):
-        container_name = f"nat-lab-dns-server-{i}-1"
-        destination_path = f"{LOG_DIR}/dns_server_{i}.log"
-
-        copy_file_from_container(container_name, "/dns-server.log", destination_path)
-
-
-def collect_core_api_server_logs():
-    container_name = "nat-lab-core-api-1"
-    os.makedirs(LOG_DIR, exist_ok=True)
-    out_path = os.path.join(LOG_DIR, "core_api.log")
-    with open(out_path, "w", encoding="utf-8") as f:
-        subprocess.run(
-            ["docker", "logs", container_name],
-            stdout=f,
-            stderr=subprocess.STDOUT,
-            text=True,
-            check=True,
-        )
-
-
-def copy_file_from_container(container_name, src_path, dst_path):
-    docker_cp_command = f"docker cp {container_name}:{src_path} {dst_path}"
-    try:
-        subprocess.run(docker_cp_command, shell=True, check=True)
-        setup_log.info(
-            "Log file %s copied successfully from %s to %s",
-            src_path,
-            container_name,
-            dst_path,
-        )
-    except subprocess.CalledProcessError:
-        setup_log.warning(
-            "Error copying log file %s from %s to %s",
-            src_path,
-            container_name,
-            dst_path,
-        )
-
-
-async def collect_mac_diagnostic_reports():
-    is_ci = "GITLAB_CI" in os.environ
-    if not (
-        is_ci
-        or "NATLAB_COLLECT_MAC_DIAGNOSTIC_LOGS" in os.environ
-        or "mac" in SESSION_VM_MARKS
-    ):
-        return
-    setup_log.info("Collect mac diagnostic reports")
-    async with SshConnection.new_connection(
-        LAN_ADDR_MAP[ConnectionTag.VM_MAC]["primary"], ConnectionTag.VM_MAC
-    ) as connection:
-        await connection.download(
-            "/Library/Logs/DiagnosticReports",
-            f"{LOG_DIR}/system_diagnostic_reports",
-        )
-        await connection.download(
-            "/root/Library/Logs/DiagnosticReports",
-            f"{LOG_DIR}/user_diagnostic_reports",
-        )
-
-
-async def start_tcpdump_processes():
-    if SESSION_SCOPE_EXIT_STACK is None:
-        raise RuntimeError("SESSION_SCOPE_EXIT_STACK is not initialized")
-    connections = []
-    for gw_tag in ConnectionTag:
-        if gw_tag in OPENWRT_VM_TAGS:
-            continue
-        if "_GW" in gw_tag.name:
-            if not await is_running(gw_tag):
-                continue
-            connection = await SESSION_SCOPE_EXIT_STACK.enter_async_context(
-                new_connection_raw(gw_tag)
-            )
-            connections.append(connection)
-    for conn_tag in [
-        ConnectionTag.DOCKER_DNS_SERVER_1,
-        ConnectionTag.DOCKER_DNS_SERVER_2,
-    ]:
-        if await is_running(conn_tag):
-            connection = await SESSION_SCOPE_EXIT_STACK.enter_async_context(
-                new_connection_raw(conn_tag)
-            )
-            connections.append(connection)
-
-    await SESSION_SCOPE_EXIT_STACK.enter_async_context(
-        make_tcpdump(connections, session=True)
-    )
-    await SESSION_SCOPE_EXIT_STACK.enter_async_context(make_local_tcpdump())
-
-
-async def start_windows_vms_resource_monitoring():
-    vms = [ConnectionTag.DOCKER_WINDOWS_VM_1, ConnectionTag.DOCKER_WINDOWS_VM_2]
-    for vm_tag in vms:
-        is_vm_running = await is_running(vm_tag)
-        if is_vm_running:
-            funcs = [
-                start_windows_vm_cpu_monitoring(vm_tag),
-                start_windows_vm_memory_monitoring(vm_tag),
-                start_windows_vm_top10_cpu_usage_monitoring(vm_tag),
-                start_windows_vm_top10_memory_usage_monitoring(vm_tag),
-            ]
-            run_windows_vm_monitoring_funcs(vm_tag, funcs)
-
-
-def run_windows_vm_monitoring_funcs(vm_tag, funcs):
-    def aux():
-        while not END_TASKS.is_set():
-            for func in funcs:
-                func()
-
-    global TASKS
-    TASKS += [
-        asyncio.create_task(asyncio.to_thread(aux))
-    ]  # Storing the task to keep it alive
-
-
-def start_windows_vm_top10_cpu_usage_monitoring(vm_tag):
-    powershell_cmd = r"""
-    Get-Counter '\Process(*)\% Processor Time' -ErrorAction SilentlyContinue |
-    Select-Object -ExpandProperty CounterSamples |
-    Where-Object {$_.InstanceName -ne '_total' -and $_.InstanceName -ne 'idle'} |
-    Sort-Object CookedValue -Descending |
-    Select-Object -First 10 InstanceName, @{Name='CPU%';Expression={[math]::Round($_.CookedValue,2)}}
-    """
-    return start_windows_vm_top10_usage_monitoring(vm_tag, "cpu", powershell_cmd)
-
-
-def start_windows_vm_top10_memory_usage_monitoring(vm_tag):
-    powershell_cmd = (
-        "ps | sort ws -desc | select -first 10 name,@{n='MB';e={[int]($_.ws/1mb)}}"
-    )
-    return start_windows_vm_top10_usage_monitoring(vm_tag, "memory", powershell_cmd)
-
-
-def start_windows_vm_top10_usage_monitoring(
-    vm_tag: ConnectionTag, resource_name: str, powershell_cmd: str
-):
-    def aux():
-        output_filename = f"logs/top10_{resource_name}_usage_{vm_tag}.txt"
-        first = True
-        with open(output_filename, "a", encoding="utf-8") as output_file:
-            # This command takes usually ~5s to complete, so I've decided not to add any
-            # additional explicit sleep
-
-            result = subprocess.run(
-                [
-                    "docker",
-                    "exec",
-                    container_id(vm_tag),
-                    "python3",
-                    "/run/qga.py",
-                    "--powershell",
-                    powershell_cmd,
-                ],
-                capture_output=True,
-                text=True,
-            )
-            if result.returncode != 0:
-                setup_log.warning(
-                    "powershell command for top10 %s failed on %s: %s",
-                    resource_name,
-                    vm_tag,
-                    result.stderr,
-                )
-                time.sleep(1)
-                return
-            lines = result.stdout.splitlines()
-            lines = list(itertools.dropwhile(lambda x: "STDOUT:" not in x, lines))[1:]
-            lines = [x.strip() for x in lines if x != ""]
-            current_time_iso = datetime.now().isoformat()
-            top10 = "\n".join(lines)
-            if first:
-                first = False
-            else:
-                output_file.write("\n")
-            output_file.write(f"{current_time_iso}{top10}\n")
-
-    return aux
-
-
-def start_windows_vm_monitoring(
-    vm_tag: ConnectionTag, resource_name: str, powershell_cmd: str
-):
-    def aux():
-        output_filename = f"logs/{resource_name}_usage_{vm_tag}.csv"
-        with open(output_filename, "a", encoding="utf-8") as output_file:
-            # This command takes usually ~5s to complete, so I've decided not to add any
-            # additional explicit sleep
-            result = subprocess.run(
-                [
-                    "docker",
-                    "exec",
-                    container_id(vm_tag),
-                    "python3",
-                    "/run/qga.py",
-                    "--powershell",
-                    powershell_cmd,
-                ],
-                capture_output=True,
-                text=True,
-            )
-            if result.returncode != 0:
-                setup_log.warning(
-                    "powershell command for %s failed on %s: %s",
-                    resource_name,
-                    vm_tag,
-                    result.stderr,
-                )
-                time.sleep(1)
-                return
-            lines = result.stdout.splitlines()
-            lines = list(itertools.dropwhile(lambda x: "STDOUT:" not in x, lines))[1:]
-            lines = [x.strip() for x in lines if x != ""]
-            current_time_iso = datetime.now().isoformat()
-            output_file.write(f"{current_time_iso}, {', '.join(lines)}\n")
-
-    return aux
-
-
-def start_windows_vm_memory_monitoring(vm_tag: ConnectionTag):
-    cmd = """
-    $os = Get-CimInstance Win32_OperatingSystem
-    $totalRAM = [math]::Round($os.TotalVisibleMemorySize/1MB, 2)
-    $freeRAM = [math]::Round($os.FreePhysicalMemory/1MB, 2)
-    $usedRAM = $totalRAM - $freeRAM
-    $percentUsed = [math]::Round(($usedRAM/$totalRAM)*100, 2)
-
-    Write-Host "$totalRAM, $usedRAM, $freeRAM, $percentUsed"
-    """
-    return start_windows_vm_monitoring(vm_tag, "memory", cmd)
-
-
-def start_windows_vm_cpu_monitoring(vm_tag: ConnectionTag):
-    cmd = "(Get-Counter '\\Processor(*)\\% Processor Time').CounterSamples.CookedValue"
-    return start_windows_vm_monitoring(vm_tag, "cpu", cmd)
+    asyncio.run(collect_logs(SESSION_VM_MARKS))
 
 
 @pytest.fixture(autouse=True)
@@ -1200,10 +252,3 @@ def setup_logger(tmp_path, request):
         yield
     finally:
         pass
-
-
-def get_required_vm_containers_from_marks():
-    containers = set()
-    for mark in SESSION_VM_MARKS:
-        containers.update(SESSION_MARK_TO_CONTAINERS.get(mark, []))
-    return containers
